requirements.txt
python-telegram-bot==20.7
PyPDF2==3.0.1
python-dotenv==1.0.0
regex==2023.12.25
pydantic==2.5.3
pydantic-settings==2.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
